Makefile: add hand-in, rework clean
The new target expects an ssh config file in the git root with the
following syntax:

Host oop-handin
    User <user>
    HostName g0.complang.tuwien.ac.at

The contents of the current directory are then rsynced to the group
directory on the hand-in server.

Additionally, reworked clean target to delete all nested class .class
files and removed the line referencing the .jar file.